dddrequirements.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
asttokens
attrs
azure-ai-inference
azure-core
beautifulsoup4
blinker
cachetools
certify
cffi
charset-normalizer
click
colorama
curl_cffi
dataclasses-json
decorator
distro
duckduckgo_search
executing
faiss-cpu
filelock
frozendict
frozenlist
fsspec
gitdb
GitPython
greenlet
h11
httpcore
httpx
httpx-sse
huggingface-hub
idna
ipython
ipython_pygments_lexers
isodate
jedi
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lxml
MarkupSafe
marshmallow
matplotlib-inline
mpmath
multidict
multitasking
mypy_extensions
narwhals
networkx
numpy
openai
orjson
ormsgpack
packaging
pandas
parso
peewee
pillow
platformdirs
primp
prompt_toolkit
propcache
protobuf
pure_eval
pyarrow
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyPDF2
python-dateutil
python-dotenv
pytz
PyYAML
referencing
regex
requests
requests-toolbelt
rpds-py
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
six
smmap
sniffio
soupsieve
SQLAlchemy
stack-data
streamlit
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
toml
torch
tornado
tqdm
traitlets
transformers
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
watchdog
wcwidth
websockets
wikipedia
xxhash
yarl
yfinance
zstandard
